video: derive the v4l2 ioctl numbers from the struct sizes
The ioctl request numbers were hard-coded for LP64, and they encode the size of
the argument struct -- which differs on ILP32: `v4l2_buffer` is 88 bytes on
64-bit and 68 on 32-bit, `v4l2_format` 208 against 204. On the 32-bit board the
driver was being handed requests it did not recognise.

A `const fn ioc(...)` now derives each number from `size_of`, so the two ABIs
follow from the struct definitions instead of from a table. `const _: () =
assert!(...)` pins the 64-bit values to what was there before, so the change is
provably a no-op on the arch that already worked.

`unsigned long` is 32-bit on armhf, so the request type, `Timeval`'s fields and
the buffer union's offset follow the pointer width, and the union carries a
zero-length `usize` array to keep its alignment.

+++ crates/yscv-video/src/framebuffer.rs
@@ -16,7 +16,10 @@ mod linux_impl {
     unsafe extern "C" {
         fn open(path: *const c_char, flags: i32) -> i32;
         fn close(fd: i32) -> i32;
-        fn ioctl(fd: i32, request: u64, ...) -> i32;
+        // `unsigned long`, so it follows the pointer width — 32-bit on armhf.
+        // Declaring it `u64` here disagreed with the same symbol in `v4l2.rs`
+        // and passed a wrong-width argument on a 32-bit board.
+        fn ioctl(fd: i32, request: usize, ...) -> i32;
         fn mmap(addr: *mut u8, len: usize, prot: i32, flags: i32, fd: i32, offset: i64) -> *mut u8;
         fn munmap(addr: *mut u8, len: usize) -> i32;
     }
@@ -28,8 +31,8 @@ mod linux_impl {
     const MAP_FAILED: *mut u8 = !0usize as *mut u8;
 
     // ioctl numbers for fbdev (from <linux/fb.h>)
-    const FBIOGET_VSCREENINFO: u64 = 0x4600;
-    const FBIOGET_FSCREENINFO: u64 = 0x4602;
+    const FBIOGET_VSCREENINFO: usize = 0x4600;
+    const FBIOGET_FSCREENINFO: usize = 0x4602;
 
     // -----------------------------------------------------------------------
     // fbdev kernel ABI structs
